refactor: устраняет указанные недочеты
