+++ src/components/header/header.tsx
@@ -1,12 +1,11 @@
 import Logo from '../logo/logo';
-import HeaderNav from './header-nav';
-import { IS_LOGGED } from '../../const';
+import HeaderNav from './components/header-nav';
 
 type HeaderProps = {
   withoutNav?: boolean;
 }
 
-export default function Header({withoutNav = false}: HeaderProps) {
+export default function Header({ withoutNav = false }: HeaderProps) {
   return (
     <header className="header">
       <div className="container">
@@ -14,8 +13,7 @@ export default function Header({withoutNav = false}: HeaderProps) {
           <div className="header__left">
             <Logo element='headerLogo' />
           </div>
-          {!withoutNav &&
-            <HeaderNav isLogged={IS_LOGGED} />}
+          {!withoutNav && <HeaderNav />}
         </div>
       </div>
     </header>
