+++ src/const.ts
@@ -1,6 +1,15 @@
 const FAVORITE_COUNT = 3;
+const FAVORITE_OFFERS_COUNT = 4;
 const OFFERS_COUNT = 5;
 const IS_LOGGED = true;
 const RATING_RATIO = 20;
+const CURRENT_CITY = 'Paris';
 
-export { FAVORITE_COUNT, OFFERS_COUNT, IS_LOGGED, RATING_RATIO };
+export {
+  FAVORITE_COUNT,
+  OFFERS_COUNT,
+  FAVORITE_OFFERS_COUNT,
+  IS_LOGGED,
+  RATING_RATIO,
+  CURRENT_CITY,
+};
