+++ src/pages/main-page/components/cities-map-section.tsx
@@ -1,0 +1,9 @@
+import { OFFERS_COUNT } from '../../../const';
+
+export default function CitiesMapSection() {
+  return (
+    <div className="cities__right-section">
+      {OFFERS_COUNT && <section className="cities__map map" />}
+    </div>
+  );
+}
