+++ src/pages/main-page/components/cities-no-places.tsx
@@ -1,0 +1,13 @@
+export default function CitiesNoPlaces() {
+  return (
+    <section className="cities__no-places">
+      <div className="cities__status-wrapper tabs__content">
+        <b className="cities__status">No places to stay available</b>
+        <p className="cities__status-description">
+          We could not find any property available at the moment in
+          Dusseldorf
+        </p>
+      </div>
+    </section>
+  );
+}
