+++ src/pages/offer-page/components/offer-map-section.tsx
@@ -1,0 +1,5 @@
+export default function OfferMapSection() {
+  return (
+    <section className="offer__map map" />
+  );
+}
